# Conflicts:
#	src/haven/Coord.java
#	src/haven/GOut.java
#	src/haven/GameUI.java
#	src/haven/Gob.java
#	src/haven/Homing.java
#	src/haven/LinMove.java
#	src/haven/LocalMiniMap.java
#	src/haven/MCache.java
#	src/haven/MapMesh.java
#	src/haven/MapMod.java
#	src/haven/MapView.java
#	src/haven/MenuGrid.java
#	src/haven/MiniMap.java
#	src/haven/Moving.java
#	src/haven/OCache.java
#	src/haven/Party.java
#	src/haven/Session.java
#	src/haven/SpriteLink.java
#	src/haven/Utils.java
#	src/haven/resutil/Ridges.java

+++ src/haven/GobPath.java
@@ -37,9 +37,9 @@ public class GobPath extends Sprite {
             return;
         Gob gob = (Gob) owner;
         Coord3f pc = gob.getc();
-        float x = lm.t.x - pc.x;
-        float y = -lm.t.y + pc.y;
-        float z = Math.sqrt(x * x + y * y) >= vd ? 0 : gob.glob.map.getcz(lm.t.x, lm.t.y) - pc.z;
+        double x = lm.t.x - pc.x;
+        double y = -lm.t.y + pc.y;
+        double z = Math.sqrt(x * x + y * y) >= vd ? 0 : gob.glob.map.getcz(lm.t.x, lm.t.y) - pc.z;
         g.apply();
         BGL gl = g.gl;
         gl.glLineWidth(2.0F);
@@ -49,7 +49,7 @@ public class GobPath extends Sprite {
         gl.glHint(GL2.GL_LINE_SMOOTH_HINT, GL2.GL_NICEST);
         gl.glBegin(GL2.GL_LINES);
         gl.glVertex3f(0, 0, 0);
-        gl.glVertex3f(x, y, z);
+        gl.glVertex3f((float)x, (float)y, (float)z); // FIXME: lossy conversion
         gl.glEnd();
         gl.glDisable(GL2.GL_LINE_SMOOTH);
     }
